Home-0.5
逻辑锁更新，但还是有问题，tablename不能传过去

+++ Yiji/master/delete_customer.asp
@@ -11,13 +11,18 @@ end if
 <!-- #include file="../conn2.asp" -->
 <%
 
+if 1=2 then
 sql="delete from customer where customername='"&request("customername")&"'"
 conn.execute(sql)
+end if
 
 '删除逻辑锁'
 sql="delete from locktable where tablename='customer' and combinedkey='"&request("customername")&"'"
 conn.execute(sql)
 
-response.redirect "master.asp"
+'response.redirect "master.asp"
 %>
 
+<script language="javascript">
+window.close();
+</script>
